# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Puraani Jeans APIs",
        "description": "Puraani Jeans API documentation",
        "contact": {
            "email": "developer@puranijeans.com"
        },
        "license": {
            "name": "Apache 2.0"
        },
        "version": ""
    },
    "servers": [
        {
            "url": "http://puranijeans.test/",
            "description": "dev"
        }
    ],
    "paths": {
...
        "/api/featured-products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get Homepage featured products",
                "operationId": "getFeaturedProducts",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "user_access_token": [],
                        "locale": []
                    }
                ]
            }
        },
        "/api/filter/featured-products": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Get Homepage featured products",
                "operationId": "getFilteredFeaturedProducts",
                "requestBody": {
                    "description": "Set Order Shipment Details",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "filters": {
                                        "description": "filters",
                                        "properties": {
                                            "price_range": {
                                                "description": "Price Range",
                                                "properties": {
                                                    "min": {
                                                        "description": "1",
                                                        "type": "string"
                                                    },
                                                    "max": {
                                                        "description": "1",
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "store_slug": {
                                                "description": "Filter by store slug",
                                                "type": "string"
                                            },
                                            "sort_by": {
                                                "description": "Sort by",
                                                "properties": {
                                                    "featured": {
                                                        "description": "1",
                                                        "type": "integer"
                                                    },
                                                    "newest_arrival": {
                                                        "description": "1",
                                                        "type": "integer"
                                                    },
                                                    "price_high_to_low": {
                                                        "description": "1",
                                                        "type": "integer"
                                                    },
                                                    "price_low_to_high": {
                                                        "description": "0",
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "user_access_token": [],
                        "locale": []
                    }
                ]
            }
        },
        "/api/recently-viewed/products": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Get customer's recently viewed products",
                "operationId": "getRecentlyViewedProducts",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "user_access_token": [],
                        "locale": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "user_access_token": {
                "type": "apiKey",
                "description": "user access token",
                "name": "Authorization",
                "in": "header"
            },
            "client_id": {
                "type": "apiKey",
                "description": "client id",
                "name": "x-client-id",
                "in": "header"
            },
            "client_secret": {
                "type": "apiKey",
                "description": "client secret",
                "name": "x-client-token",
                "in": "header"
            }
        }
    }
}